Fixed warning
git-svn-id: file:///home/michael/svn/yuni/trunk@1368 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/file/file.cpp
@@ -215,7 +215,7 @@ namespace IO
 			if (!fstat(fdIN, &st))
 			{
 				off_t offset = 0;
-				if (-1 != sendfile(fdOUT, fdIN, &offset, st.st_size))
+				if (-1 != sendfile(fdOUT, fdIN, &offset, (size_t) st.st_size))
 					return Yuni::Core::IO::ioErrNone;
 			}
